search test convert to junit
